requirements.txt
# Windows Runtime packages for notification listening
winrt-runtime>=2.0.0
winrt-Windows.UI.Notifications>=2.0.0
winrt-Windows.Foundation>=2.0.0
winrt-Windows.Foundation.Collections>=2.0.0

# HTTP client for provider APIs (Pushbullet, Ntfy)
requests>=2.31.0

# Environment configuration
python-dotenv>=1.0.0

# Firebase Cloud Messaging authentication
google-auth>=2.23.0
